# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Starter Kit API
  version: 1.0.0
  description: REST API for the Starter Kit application
servers:
  - url: /api
    description: Default server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly access token cookie set on login
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        organizationId:
          type: string
          format: uuid
          example: 8b6b3f2e-1c1a-4a2b-9c3d-0f6a2e8b7a10
        email:
          type: string
          format: email
          example: user@example.com
        fullName:
          type: string
          example: Jane Doe
        role:
          type: string
          description: >
            OWNER and ADMIN both have full User Management access; INTERNAL
            is read-only. OWNER cannot be assigned via the role-change
            endpoint — it is only ever set through organization ownership
            assignment/transfer.
          enum: [OWNER, ADMIN, INTERNAL]
          example: ADMIN
        status:
          type: string
          enum: [INVITED, ACTIVE, SUSPENDED, DEACTIVATED]
          example: ACTIVE
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-01T00:00:00.000Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 7c8b9a2e-4d3f-4e5a-8b6c-7d8e9f0a1b2c
        email:
          type: string
          format: email
          example: new.hire@acmelegal.com
        role:
          type: string
          enum: [ADMIN, INTERNAL]
          example: INTERNAL
        status:
          type: string
          enum: [PENDING, ACCEPTED, EXPIRED, REVOKED]
          example: PENDING
        expiresAt:
          type: string
          format: date-time
          example: "2024-01-08T00:00:00.000Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    CreateInvitationBody:
      type: object
      required: [email, fullName, role]
      properties:
        email:
          type: string
          format: email
          example: new.hire@acmelegal.com
        fullName:
          type: string
          minLength: 2
          maxLength: 100
          example: Jordan Lee
        role:
          type: string
          enum: [ADMIN, INTERNAL]
          example: INTERNAL
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
...
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Invalid email address
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 87
        totalPages:
          type: integer
          example: 5

    # ─── Contract Management System ────────────────────────────────────────────
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users in the current organization
      operationId: listUsers
      description: >
        Available to any authenticated member of the organization — viewing
        the roster is not admin-only, only the mutating endpoints below are.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /invitations:
    post:
      tags: [Users]
      summary: Invite a new user
      operationId: createInvitation
      description: >
        Requires the `OWNER` or `ADMIN` role. Creates a pending Invitation
        and emails an accept-invitation link; no User row is created until
        the invitation is accepted (see `POST /auth/accept-invitation`).
        `role` cannot be `OWNER`.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Invitation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A user or pending invitation already exists for this email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"

    get:
      tags: [Users]
      summary: List pending invitations for the current organization
      operationId: listInvitations
      description: >
        Available to any authenticated member of the organization, matching
        `GET /users`. Creating or resending invitations is admin-only.
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Paginated list of pending invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
                  meta:
                    type: object
                    properties:
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invitations/{id}/resend:
    post:
      tags: [Users]
      summary: Resend a pending invitation
      operationId: resendInvitation
      description: >
        Requires the `OWNER` or `ADMIN` role. Issues a new token and
        expiry, invalidating the previous one, and re-sends the invitation
        email. Also revives an `EXPIRED` invitation back to `PENDING`.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          example: 7c8b9a2e-4d3f-4e5a-8b6c-7d8e9f0a1b2c
      responses:
        "200":
          description: Invitation resent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Invitation"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Invitation already accepted or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...